+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+### Fixed
+- agent: Import of ClusterShell NodeSet class (#682). Contribution from
+  @faganihajizada.
+
 ## [6.0.0] - 2025-11-28
 
 ### Added